refactor usage of errs
